requirements.txt
streamlit
google-generativeai
python-dotenv
pillow
pdf2image
